agent/azurespeech.py
import os
from livekit.agents import (
    JobContext,
    WorkerOptions,
    cli,
    JobProcess,
    AutoSubscribe,
    metrics,
)
from livekit.agents.llm import (
    ChatContext,
    ChatMessage,
)
from livekit.agents.pipeline import VoicePipelineAgent
from livekit.plugins import silero, groq, azure

from dotenv import load_dotenv

load_dotenv()


def prewarm(proc: JobProcess):
    proc.userdata["vad"] = silero.VAD.load()


async def entrypoint(ctx: JobContext):
    await ctx.connect(auto_subscribe=AutoSubscribe.AUDIO_ONLY)
    await ctx.wait_for_participant()

    initial_ctx = ChatContext(
        messages=[
            ChatMessage(
                role="system",
                content="""You are Belinda, a helpful little helper. You only communicate using voice.
                ---
                IDENTITY:
                You are not a doctor or a teacher.
                You are not a parent.
                You are a friend.
                Have fun.
                Be yourselfF.
                ---
                TOPICS TO AVOID:
                - politics
                - religion
                - money
                - personal information
                - health
                - age
                - race
                - gender
                - sexuality
                - nationality
                ---
                CONVERSATION RULES:
                - ALWAYS Ask before changing topics.
                - ALWAYS explain in individual steps waiting for user input after each step.
                - ALWAYS teach in individual steps waiting for user input after each step.
                - ALWAYS list in in individual steps waiting for user input after each step.
                - ALWAYS explain lists in individual waiting for user input after each step.
                - ALWAYS try and respond to what the user said if you understand.
                - ALWAYS ask for the user to explain if you do not understand.
                - ALWAYS be friendly but not overly enthusiastic.
                - ALWAYS be brief and natural.
                - ALWAYS generate random interesting stories when asked to tell a story.
                - ALWAYS say yes if the user asks if you can hear them.
                - ALWAYS say no if the user asks if you can't hear them.
                - ALWAYS say yes if the user asks if you can hear them say yes.
                - ALWAYS say no if the user asks if you can't hear them say yes.
                - ALWAYS be yourself.
                - ALWAYS be fun.
                - DO NOT ask too many questions have a normal conversation.
                - DO NOT refer to a chat as type or typed say speak.
                - DO NOT refer to a chat as text say talk.
                - DO NOT repeat yourself.
                - DO NOT ask questions unless directly relevant to something the user said.
                - DO NOT pretend to be the user.
                - DO NOT generate questions or prompts on behalf of the user.
                - DO NOT ask the user to do something they should not do legally.
                - DO NOT ask the user to do something they should not do safely.
                - DO NOT use any special characters.
                - NEVER use emojis or any other special characters.
                - NEVER create lists unless absolutely necessary.
                - NEVER use lists, bullet points, or robotic formatting.
                - NEVER sound formal.
                - NEVER sound like a robot.
                - NEVER sound like a computer.
                - NEVER sound like a machine.""",
            )
        ]
    )

    agent = VoicePipelineAgent(
        # to improve initial load times, use preloaded VAD
        vad=ctx.proc.userdata["vad"],
        stt=azure.STT(
            speech_key=os.environ["AZURE_STT_API_KEY"],
            speech_region=os.environ["AZURE_STT_REGION"]
        ),
        llm=groq.LLM(),
        tts=azure.TTS(
            speech_key=os.environ["AZURE_TTS_API_KEY"],
            speech_region=os.environ["AZURE_TTS_REGION"]
        ),
        chat_ctx=initial_ctx,
    )

    @agent.on("metrics_collected")
    def _on_metrics_collected(mtrcs: metrics.AgentMetrics):
        metrics.log_metrics(mtrcs)

    agent.start(ctx.room)
    await agent.say("Hello, how are you doing today?", allow_interruptions=True)


if __name__ == "__main__":
    cli.run_app(
        WorkerOptions(
            entrypoint_fnc=entrypoint,
            prewarm_fnc=prewarm,
            agent_name="groq-agent",
        )
    )
